requirements.txt
ping3
rich
Flask
scapy
requests
PyYAML
